Move rest of payload utils to live_image utils
All the functions there are live_image payload specific utils.
