+++ pyanaconda/modules/payload/payloads/live_image/utils.py
